IE compatibility and split code into more modules
